Implement marshalling plus conversion of JavaScript objects into Java
This allows us to inject complex types directly onto a Java class. The
implementation relies on Jackson and it's corresponding annotations to
handle and control the conversion.

This also moves JSParamActionListener into a separate class file - making
the actual extension slightly cleaner.
